Reorganized token trade modal components
