+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/CommonTradeModal.scss
@@ -1,6 +1,6 @@
-@import '../../../styles/shared.scss';
+@import '../../../../styles/shared.scss';
 
-.TradeTokenModal {
+.CommonTradeModal {
   overflow-y: scroll;
   padding-left: 8px;
