+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TokenIcon/TokenIcon.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared.scss';
+@import '../../../../../styles/shared.scss';
 
 .TokenIcon {
   border-radius: 50%;
