+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TokenIcon/TokenIcon.tsx
