+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TokenIcon/index.ts
