+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm/AddressBalance/AddressBalance.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared.scss';
+@import '../../../../../../styles/shared.scss';
 
 .AddressBalance {
   display: flex;
