+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm/AmountSelections/AmountSelections.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared.scss';
+@import '../../../../../../styles/shared.scss';
 
 .AmountSelections {
   display: flex;
