+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm/ReceiptDetails/BuyReceipt.tsx
