+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm/ReceiptDetails/ReceiptDetails.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared.scss';
+@import '../../../../../../styles/shared.scss';
 
 .ReceiptDetails {
   display: flex;
