+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm/ReceiptDetails/SellReceipt.tsx
