+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm/TradeTokenForm.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared.scss';
+@import '../../../../../styles/shared.scss';
 
 .TradeTokenForm {
   display: flex;
