+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/index.ts
@@ -1,0 +1,3 @@
+import CommonTradeModal from './CommonTradeModal';
+
+export default CommonTradeModal;
