+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/types.ts
@@ -1,0 +1,7 @@
+import { TradingConfig } from './CommonTradeModal/TradeTokenForm';
+
+export type TradeTokenModalProps = {
+  isOpen: boolean;
+  onModalClose?: () => void;
+  tradeConfig: TradingConfig;
+};
